+++ internal/api/functions.go
@@ -1,7 +1,6 @@
 package api
 
 import (
-	"database/sql"
 	"encoding/json"
 	"errors"
 	"log"
@@ -9,8 +8,6 @@ import (
 	"strconv"
 	"strings"
 	"time"
-
-	"github.com/igortoigildin/todo_app/internal/dbs"
 )
 
 func NextDate(now time.Time, date string, repeat string) (string, error) {
@@ -77,33 +74,6 @@ func formatDate(date string) (string, error) {
 	dateFormatted := result.Format("20060102")
 	return dateFormatted, nil
 }
-
-func sendTaskToDB(w http.ResponseWriter, task Task) (IdStrusct, error) {
-	var taskId IdStrusct
-	// open and check db connection 
-	db, err := dbs.ConnectDB("scheduler.db")
-	if err != nil {
-		log.Fatalf("unable to connect to database: %v", err)
-	}
-	// sending received task to db
-	res, err := db.Exec("INSERT INTO scheduler (date, comment, title, repeat) VALUES (:date, :comment, :title, :repeat)",
-	sql.Named("date", task.Date),
-	sql.Named("comment", task.Comment),
-	sql.Named("title", task.Title),
-	sql.Named("repeat", task.Repeat))
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusBadRequest)
-		return taskId, err
-	}
-	// getting the last inserted task
-	id, err := res.LastInsertId()
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusBadRequest)
-		return taskId, err
-	}
-	taskId.Id = id
-	return taskId, nil
-}
 // creating method for error handling in JSON
 func JSONError(w http.ResponseWriter, err interface{}, code int) {
     w.Header().Set("Content-Type", "application/json; charset=utf-8")
@@ -114,34 +84,3 @@ func JSONError(w http.ResponseWriter, err interface{}, code int) {
     json.NewEncoder(w).Encode(result)
 }
 
-func checkIfTaskRequestValid(w http.ResponseWriter, task Task) bool {
-	// check if title line is empty
-	if task.Title == "" {
-		JSONError(w, "не указан заголовок задачи", http.StatusBadRequest)
-		return false
-	}
-	// check if time format is valid
-	dateReceived, err := time.Parse("20060102", task.Date)
-	if err != nil {
-		JSONError(w, "Дата представлена в некорректном формате", http.StatusBadRequest)
-		return false
-	}
-	// check repeat format
-	timeNow, err := time.Parse("20060102", currentDate())
-	if err != nil {
-		JSONError(w, "Interanal server error", http.StatusInternalServerError)
-		return false
-	}
-	// check if repeat format is valid and if date received before current date
-	if task.Repeat != "" && dateReceived.Unix() < timeNow.Unix() {
-		nextDate, err := NextDate(timeNow, task.Date, task.Repeat)
-		if err != nil {
-			JSONError(w, "Repeat format is not valid", http.StatusBadRequest)
-			return false
-		}
-		task.Date = nextDate
-	} else if task.Repeat == "" && dateReceived.Unix() < timeNow.Unix() {
-		task.Date = currentDate()
-	}
-	return true
-}
